+++ adapters/cx.pl
@@ -251,7 +251,6 @@ setup_call_cleanup(Setup, Call, Cleanup) :-
 	).
 '$lgt_default_flag'(report, on).
 '$lgt_default_flag'(clean, on).
-'$lgt_default_flag'(reload, always).
 '$lgt_default_flag'(code_prefix, '$').
 '$lgt_default_flag'(optimize, on).
 '$lgt_default_flag'(source_data, on).