+++ adapters/eclipse.pl
@@ -253,7 +253,6 @@ setup_call_cleanup(_, _, _) :-
 	).
 '$lgt_default_flag'(report, on).
 '$lgt_default_flag'(clean, on).
-'$lgt_default_flag'(reload, always).
 '$lgt_default_flag'(code_prefix, '$').
 '$lgt_default_flag'(optimize, on).
 '$lgt_default_flag'(source_data, on).