+++ adapters/qp.pl
@@ -282,7 +282,6 @@ format(Format, Arguments) :-
 '$lgt_default_flag'(scratch_directory, '.lgt_tmp/').
 '$lgt_default_flag'(report, on).
 '$lgt_default_flag'(clean, on).
-'$lgt_default_flag'(reload, always).
 '$lgt_default_flag'(code_prefix, '$').
 '$lgt_default_flag'(optimize, on).
 '$lgt_default_flag'(source_data, on).