+++ adapters/unsupported/eclipse6.pl
@@ -345,7 +345,6 @@ call(F, A1, A2, A3, A4, A5, A6) :-
 	).
 '$lgt_default_flag'(report, on).
 '$lgt_default_flag'(clean, on).
-'$lgt_default_flag'(reload, always).
 '$lgt_default_flag'(code_prefix, '$').
 '$lgt_default_flag'(optimize, on).
 '$lgt_default_flag'(source_data, on).