+++ adapters/xsb.pl
@@ -238,7 +238,6 @@ setup_call_catcher_cleanup(Setup, Call, Catcher, Cleanup) :-
 	).
 '$lgt_default_flag'(report, on).
 '$lgt_default_flag'(clean, on).
-'$lgt_default_flag'(reload, always).
 '$lgt_default_flag'(code_prefix, '$').
 '$lgt_default_flag'(optimize, on).
 '$lgt_default_flag'(source_data, on).